refactor : baseballnumber과 hint의 역할 분리
