+++ src/main/java/baseball/BaseballNumber.java
@@ -7,6 +7,9 @@ public class BaseballNumber {
 
     public static final int NUMBER_OF_CASE = 3;
 
+    public static final int BALL = 1;
+    public static final int NOTBALL = 0;
+
     public void pickNum(List<Integer> randomNum) {
         while (randomNum.size() != NUMBER_OF_CASE) {
             int ranNum = Randoms.pickNumberInRange(1, 9);
@@ -16,20 +19,32 @@ public class BaseballNumber {
         }
     }
 
-    public List<Integer> getUserNum() {
-        System.out.printf("숫자를 입력해주세요");
-        String[] userArr = Console.readLine().split("");
-        checkUserNum(userArr);
-        List<Integer> userNum = new ArrayList<>();
-        userNum.add(Integer.parseInt(userArr[0]));
-        userNum.add(Integer.parseInt(userArr[1]));
-        userNum.add(Integer.parseInt(userArr[2]));
-        return userNum;
+    public int countStrike(List<Integer> randomNum, List<Integer> userNum) {
+        int cnt = 0;
+        for (int i = 0; i < NUMBER_OF_CASE; i++) {
+            if (randomNum.get(i) == userNum.get(i)) {
+                cnt++;
+            }
+        }
+        return cnt;
     }
 
-    public void checkUserNum(String[] userArr) {
-        if (userArr.length != NUMBER_OF_CASE) {
-            throw new IllegalArgumentException();
+    public int countBall(List<Integer> userNum, List<Integer> randomNum) {
+        int cnt = 0;
+        for (int i = 0; i < NUMBER_OF_CASE; i++) {
+            if (checkBall(i, userNum, randomNum) == BALL) {
+                cnt++;
+            }
         }
+        return cnt;
+    }
+
+    public int checkBall(int index, List<Integer> userNum, List<Integer> randomNum) {
+        if (randomNum.contains(userNum.get(index))) {
+            if (index != randomNum.indexOf(userNum.get(index))) {
+                return BALL;
+            }
+        }
+        return NOTBALL;
     }
 }
